extract `duplicate_underscore_argument`, and move it into `functions`
misc: use `str::starts_with`

extract `duplicate_underscore_argument` into a module

move it to `functions`
